Improvements to block content naming and file structure
