+++ packages/core/src/extensions/Blocks/index.ts
@@ -1,9 +1,9 @@
 import { Node } from "@tiptap/core";
 import { Block } from "./nodes/Block";
 import { BlockGroup } from "./nodes/BlockGroup";
-import { TextContent } from "./nodes/TextContent";
-import { HeadingContent } from "./nodes/HeadingContent";
-import { ListItemContent } from "./nodes/ListItemContent";
+import { TextContent } from "./nodes/BlockTypes/TextBlock/TextContent";
+import { HeadingContent } from "./nodes/BlockTypes/HeadingBlock/HeadingContent";
+import { ListItemContent } from "./nodes/BlockTypes/ListItemBlock/ListItemContent";
 
 export const blocks: any[] = [
   TextContent,
