+++ packages/core/src/extensions/Blocks/nodes/BlockTypes/TextBlock/TextContent.ts
@@ -1,5 +1,5 @@
 import { Node } from "@tiptap/core";
-import styles from "./Block.module.css";
+import styles from "../../Block.module.css";
 
 export const TextContent = Node.create({
   name: "textContent",
@@ -10,7 +10,6 @@ export const TextContent = Node.create({
     return [
       {
         tag: "p",
-        priority: 100,
       },
     ];
   },
@@ -19,9 +18,8 @@ export const TextContent = Node.create({
     return [
       "div",
       {
-        "data-node-type": "block-content", // TODO: only for testing? if so, rename to data-test-*?
-        "data-content-type": this.name,
         class: styles.blockContent,
+        "data-content-type": this.name,
       },
       ["p", 0],
     ];
